Move DATA to autoloading class
`DATA` was previously only used within `data.php` to create the `$DATA`
object. This is now done in `init.php`, and the new `Data` class is
autoloaded.
